education-ai-suite/smart-classroom/content_search/requirements_providers.txt
--extra-index-url https://download.pytorch.org/whl/cpu

chromadb==1.5.5
minio==7.2.20

fastapi==0.121.3
fastapi-utils==0.8.0
uvicorn==0.38.0

pydantic==2.11.7
pydantic-settings==2.10.1
typing-inspect==0.9.0

openvino==2025.3.0
openvino-tokenizers==2025.3.0.0
openvino-genai==2025.3.0.0
optimum-intel==1.26.0

torch==2.8.0
torchvision==0.23.0
transformers==4.53.3

llama-index==0.14.19
llama-index-embeddings-openvino==0.7.0
llama-index-readers-file>=0.4.0

# Inlined CLIP embedding (replaces multimodal_embedding_serving wheel)
open-clip-torch>=3.1.0
timm>=1.0.19

qwen-vl-utils==0.0.11
aiohttp==3.13.2

moviepy>=2.0
# NOTE: no official Windows wheel for Python 3.12;
#       install from https://github.com/dmlc/decord/releases or via conda
decord
opencv-python-headless==4.10.0.84

# NOTE: pdf2image requires Poppler binaries on Windows: https://github.com/oschwartz10612/poppler-windows/releases
pdf2image==1.17.0
# NOTE: pytesseract requires Tesseract OCR binary on Windows: https://github.com/UB-Mannheim/tesseract/wiki
pytesseract==0.3.13
# NOTE: unstructured[all-docs] has heavy system-level dependencies (detectron2, libmagic, etc.)
unstructured[all-docs]==0.18.32
python-docx==1.2.0
docx2txt==0.9

requests==2.32.4
tqdm==4.67.0
numpy>=1.24
pillow>=11.0.0
PyYAML==6.0.2

pytest==8.3.5
